+++ CHANGELOG.md
@@ -30,6 +30,7 @@ All notable changes to this project will be documented in this file.
 - x/y coordinate swap in photometry
 - lamp.nearfield parameter returning float and not bool
 - lamp.irradiance_at() function not using lamp.intensity_factor properly
+- break ties in downlight placement by preferring centroid
 
 ### Changed